+++ Comp2023/src/main/java/frc/robot/lib/util/PhoenixUtil6.java
@@ -1,4 +1,4 @@
-package frc.robot.team2135;
+package frc.robot.lib.util;
 
 import com.ctre.phoenix6.StatusCode;
 import com.ctre.phoenix6.StatusSignal;
