+++ Comp2023/src/main/java/frc/robot/subsystems/Wrist.java
@@ -38,7 +38,7 @@ import frc.robot.Robot;
 import frc.robot.RobotContainer;
 import frc.robot.lib.math.Conversions;
 import frc.robot.lib.util.CTREConfigs6;
-import frc.robot.team2135.PhoenixUtil6;
+import frc.robot.lib.util.PhoenixUtil6;
 
 //
 // Wrist subsystem class
